pyproject.toml
[project]
name = "muzharpg-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.33.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.2.2",
    "mcp>=1.27.0",
]
